requirements.txt
selenium==4.1.3
flask==2.1.1
undetected-chromedriver==3.1.5.post4
requests==2.28.1
webdriver-manager==3.8.2
Fabric3==1.14.post1
Jinja2==3.1.2
gunicorn==20.1.0
